MODULO02/DIA07/requirements.txt
mysql-connector-python
tabulate 